Delete old fretboard component
