+++ package.json
@@ -66,10 +66,8 @@
     "howler": "^2.1.1",
     "lodash": "^4.17.11",
     "react": "^16.8.4",
-    "react-anime": "^2.2.0",
     "react-dom": "^16.8.4",
     "react-scripts": "2.1.8",
-    "react-spring": "^8.0.19",
     "rebass": "^3.0.1",
     "redux": "^4.0.1",
     "redux-localstorage-simple": "^2.1.6",
