+++ src/components/Fretboard/Fretboard.tsx
@@ -1,136 +1,160 @@
 import React from "react"
-import styled from "styled-components"
-import { Box, Image, Flex } from "rebass"
+import styled, { css } from "styled-components"
+import { Box, Flex } from "rebass"
 import { isEqual } from "lodash"
 
 import fretboardGraphic from "src/assets/fretboard.jpg"
+import { notes, Note as NoteProps } from "src/utils/fretboardUtils"
 import { Display } from "src/components/ui/Typography"
 import { useStore } from "src/utils/hooks"
-import { notes, containsSharpOrFlat, getNote } from "src/utils/fretboardUtils"
-import { color } from "src/Theme"
+import { getNote } from "src/utils/fretboardUtils"
 
-export const Fretboard = _props => {
-  const { accidentalMode, showHint, showNotes } = useStore(state => state.settings) // prettier-ignore
-  const { currentNote } = useStore(state => state.notes)
+interface FretboardProps {
+  selectedNotes?: NoteProps[]
+  isVisible?: (props?) => boolean
+  renderNote?: (props?) => React.ReactNode
+}
 
+export const Fretboard: React.FC<FretboardProps> = props => {
+  const { selectedNotes, isVisible = () => true } = props
+  const { accidentalMode } = useStore(state => state.settings)
   const fretboard = notes[accidentalMode]
 
   return (
-    <Container flexDirection="column" justifyContent="center">
-      <Image width="100%" height={260} src={fretboardGraphic} />
+    <FretboardContainer>
+      {fretboard.map((string, stringIndex) => {
+        const updatePosition = computeNotePosition()
 
-      <NotesContainer ml={-40}>
-        {fretboard.map((string, stringIndex) => {
-          return (
-            <Flex key={stringIndex}>
-              {string.map((note, noteIndex) => {
-                const BASE = 98
-                const DISTANCE_RATIO = 0.65
-                const SPACE = BASE - (BASE / 12) * (noteIndex * DISTANCE_RATIO)
-                const isCurrentNote = isEqual(currentNote.position, [
-                  stringIndex,
-                  noteIndex,
-                ])
+        return (
+          <Flex key={stringIndex}>
+            {string.map((note, noteIndex) => {
+              const notePosition = updatePosition(stringIndex, noteIndex)
 
-                const getVisibility = () => {
-                  switch (true) {
-                    case showNotes: {
-                      if (accidentalMode === "naturals") {
-                        if (containsSharpOrFlat(note)) {
-                          return false
-                        }
-                      }
-                      return true
-                    }
-                    case isCurrentNote && showHint:
-                      return true
-                    default:
-                      return false
+              let currentNote
+              if (selectedNotes) {
+                currentNote = selectedNotes.find(n => {
+                  const match = isEqual([stringIndex, noteIndex], n.position)
+                  return match
+                })
+              }
+
+              const visible = isVisible({
+                ...props,
+                currentNote,
+              })
+
+              return (
+                <NoteContainer
+                  key={noteIndex}
+                  style={notePosition}
+                  onClick={() =>
+                    logNote({ stringIndex, noteIndex, accidentalMode })
                   }
-                }
-
-                // Natural notes are falsy
-                const showNote = Boolean(note) && getVisibility()
-
-                return (
-                  <NoteContainer
-                    mr={SPACE}
-                    key={noteIndex}
-                    isCurrentNote={showNote || isCurrentNote}
-                    isRoot={!showHint && showNote && note === currentNote.note}
-                    showNote={showNote}
-                  >
-                    <Note
-                      size="5"
-                      showNote={showNote}
-                      onClick={() => {
-                        const noteLookup: any = [stringIndex + 1, noteIndex]
-                        const found = getNote({
-                          accidentalMode,
-                          position: noteLookup,
-                        })
-
-                        console.warn(`[${noteLookup}]`, found)
-                      }}
-                    >
-                      {note}
+                >
+                  {props.renderNote ? (
+                    props.renderNote({
+                      ...props,
+                      Note,
+                      currentNote,
+                      note,
+                      stringIndex,
+                      noteIndex,
+                      visible,
+                    })
+                  ) : (
+                    <Note selected={Boolean(currentNote)} visible={visible}>
+                      <Display>{note}</Display>
                     </Note>
-                  </NoteContainer>
-                )
-              })}
-            </Flex>
-          )
-        })}
-      </NotesContainer>
-    </Container>
+                  )}
+                </NoteContainer>
+              )
+            })}
+          </Flex>
+        )
+      })}
+    </FretboardContainer>
   )
 }
 
-const Container = styled(Flex)`
-  user-select: none;
-  align-content: center;
+const FretboardContainer = styled(Box)`
+  background: url(${fretboardGraphic}) no-repeat;
+  background-size: 100% 100%;
+  height: 260px;
   position: relative;
 `
 
-const NotesContainer = styled(Box)`
-  position: absolute;
-`
-
-const Note = styled(Display)<{ showNote: boolean }>`
-  position: relative;
-  left: 8px;
-  color: ${color("white")};
-  visibility: ${p => (p.showNote ? "visible" : "hidden")};
-`
-
-const NoteContainer = styled(Flex)<{
-  isCurrentNote: boolean
-  isRoot?: boolean
-  showNote: boolean
-}>`
-  border-radius: 50%;
-  background-color: ${p =>
-    p.showNote ? "rgba(255, 255, 255, .1)" : "rgba(243, 251, 81, .8)"};
-
-  background-color: ${p => p.isRoot && "red"};
-
+const noteSize = css`
   width: 30px;
   height: 30px;
-  text-shadow: 4px 4px 6px rgba(0, 0, 0, 0.6);
-  box-shadow: 2px 2px 2px rgba(0, 0, 0, 0.3);
-  top: 5px;
-  margin-bottom: 10px;
-  position: relative;
-  visibility: ${p => (p.isCurrentNote ? "visible" : "hidden")};
+`
 
-  animation-name: ${p => (p.isCurrentNote ? "fadeInNote" : "none")};
+const NoteContainer = styled(Box)`
+  position: absolute;
+  ${noteSize};
+`
+
+export interface FretboardNoteProps {
+  isRoot?: boolean
+  selected: boolean
+  visible?: boolean
+  children: React.ReactNode
+}
+
+const Note = styled(Flex)<FretboardNoteProps>`
+  border-radius: 50%;
+
+  ${props => {
+    switch (true) {
+      case props.isRoot:
+        return css`
+          background-color: red;
+        `
+      case props.visible:
+        return css`
+          background-color: rgba(255, 255, 255, 0.1);
+        `
+      default:
+        return css`
+          background-color: rgba(243, 251, 81, 0.8);
+        `
+    }
+  }};
+
+  ${props => {
+    switch (true) {
+      case props.visible || props.selected:
+        return css`
+          visibility: visible;
+        `
+      default:
+        return css`
+          visibility: hidden;
+        `
+    }
+  }}
+
+  color: white;
+
+  ${noteSize};
+
+  align-items: center;
+  justify-content: center;
+  position: absolute;
+
+  box-shadow: 2px 2px 2px rgba(0, 0, 0, 0.3);
+  text-shadow: 4px 4px 6px rgba(0, 0, 0, 0.6);
+
+  /* TODO: Move this animation out of CSS */
+  animation-name: ${p => (p.visible || p.selected ? "fadeInNote" : "none")};
   animation-duration: 0.5s;
   animation-fill-mode: both;
   animation-timing-function: cubic-bezier(0.68, -0.55, 0.265, 1.85);
 
+  /* Start animation */
   opacity: 0;
   transform: scale(0.5);
 
+  /* Transition to */
   @keyframes fadeInNote {
     0% {
       opacity: 0;
@@ -138,7 +162,37 @@ const NoteContainer = styled(Flex)<{
 
     100% {
       opacity: 1;
-      transform: scale(${p => (p.showNote ? "1" : "1")});
+      transform: scale(${p => (p.visible ? "1" : "1")});
     }
   }
 `
+
+// Helpers
+
+/**
+ * TODO: Make this dynamic based upon the width of the fretboard, or find some
+ * proper fretboard fret spacing math...
+ */
+function computeNotePosition(lastPos: number = 0) {
+  return (stringIndex: number, noteIndex: number) => {
+    const base = 135
+    let left = base + lastPos - (base / 12) * (noteIndex * 0.5)
+    lastPos = left
+    left -= 167
+    const top = 15 + stringIndex * 40
+    return {
+      left,
+      top,
+    }
+  }
+}
+
+function logNote({ stringIndex, noteIndex, accidentalMode }) {
+  const noteLookup: any = [stringIndex + 1, noteIndex]
+  const found = getNote({
+    accidentalMode,
+    position: noteLookup,
+  })
+
+  console.warn(`[${noteLookup}]`, found)
+}
